# openapi/definicion.yaml
openapi: 3.1.0
info:
  title: CatalogIA API
  description: FastAPI backend for CatalogIA platform with AI recommendations
  version: 1.0.0
  contact:
    name: Equipo 22
    email: 'equipo22@tec.mx'
    url: http://localhost:8000
servers:
- url: http://localhost:8000
  description: Local development server
paths:
  /api/v1/users/register:
    post:
      tags:
      - users
      summary: Create User
      description: "Crear un nuevo usuario en el sistema.\n\nArgs:\n    `db`: Sesión\
        \ de base de datos\n    `user_in`: Datos del usuario a crear\n\nReturns:\n\
        \    `UserResponse`: Usuario creado con todos sus datos\n    \nRaises:\n \
        \   `HTTPException`: 400 si ya existe un usuario con el mismo email o username"
      operationId: create_user_api_v1_users_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/login:
    post:
      tags:
      - users
      summary: Login For Access Token
      description: "Iniciar sesión y obtener token de acceso JWT.\n\nArgs:\n    `db`:\
        \ Sesión de base de datos\n    `user_in`: Credenciales de login (email y password)\n\
        \nReturns:\n    `Token`: Token de acceso JWT y tipo de token\n    \nRaises:\n\
        \    `HTTPException`: 401 si las credenciales son incorrectas"
      operationId: login_for_access_token_api_v1_users_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/me:
    get:
      tags:
      - users
      summary: Read User Me
      description: "Obtener información del usuario autenticado actual.\n\nArgs:\n\
        \    `current_user`: Usuario autenticado mediante token JWT\n\nReturns:\n\
        \    `UserResponse`: Datos completos del usuario actual"
      operationId: read_user_me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - HTTPBearer: []
  /api/v1/users/{user_id}:
    get:
      tags:
      - users
      summary: Read User
      description: "Obtener un usuario específico por su ID.\n\nArgs:\n    `user_id`:\
        \ ID del usuario a buscar\n    `db`: Sesión de base de datos\n\nReturns:\n\
        \    `UserResponse`: Datos completos del usuario\n\nRaises:\n    `HTTPException`:\
        \ 404 si el usuario no existe"
      operationId: read_user_api_v1_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - users
      summary: Update User
      description: "Actualizar los datos de un usuario.\n\nArgs:\n    `db`: Sesión\
        \ de base de datos\n    `user_id`: ID del usuario a actualizar\n    `user_in`:\
        \ Datos del usuario a actualizar\n    `current_user`: Usuario autenticado\
        \ actual\n\nReturns:\n    `UserResponse`: Usuario actualizado con los nuevos\
        \ datos\n\nRaises:\n    `HTTPException`: 404 si el usuario no existe\n   \
        \ `HTTPException`: 403 si no tienes permisos para actualizar el usuario"
      operationId: update_user_api_v1_users__user_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - users
      summary: Delete User
      description: "Eliminar un usuario del sistema.\n\nArgs:\n    `db`: Sesión de\
        \ base de datos\n    `user_id`: ID del usuario a eliminar\n    `current_user`:\
        \ Usuario autenticado actual\n\nReturns:\n    `dict`: Mensaje de confirmación\
        \ de eliminación\n\nRaises:\n    `HTTPException`: 404 si el usuario no existe\n\
        \    `HTTPException`: 403 si no tienes permisos para eliminar el usuario"
      operationId: delete_user_api_v1_users__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/wishlist/:
    get:
      tags:
      - wishlist
      summary: Get User Wishlist
      description: "Obtener la lista de deseos completa del usuario autenticado.\n\
        \nArgs:\n    `db`: Sesión de base de datos\n    `current_user`: Usuario autenticado\
        \ actual\n\nReturns:\n    `WishlistList`: Lista completa de productos en la\
        \ wishlist con detalles del producto"
      operationId: get_user_wishlist_api_v1_wishlist__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistList'
      security:
      - HTTPBearer: []
    post:
      tags:
      - wishlist
      summary: Add To Wishlist
      description: "Agregar un producto a la lista de deseos del usuario.\n\nArgs:\n\
        \    `db`: Sesión de base de datos\n    `wishlist_in`: Datos del producto\
        \ a agregar\n    `current_user`: Usuario autenticado actual\n\nReturns:\n\
        \    `WishlistResponse`: Item agregado a la wishlist\n    \nRaises:\n    `HTTPException`:\
        \ 404 si el producto no existe\n    `HTTPException`: 400 si el producto ya\
        \ está en la wishlist"
      operationId: add_to_wishlist_api_v1_wishlist__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WishlistCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/wishlist/{product_id}:
    delete:
      tags:
      - wishlist
      summary: Remove From Wishlist
      description: "Eliminar un producto de la lista de deseos del usuario.\n\nArgs:\n\
        \    `db`: Sesión de base de datos\n    `product_id`: UUID del producto a\
        \ eliminar de la wishlist\n    `current_user`: Usuario autenticado actual\n\
        \nReturns:\n    `dict`: Mensaje de confirmación de eliminación\n    \nRaises:\n\
        \    `HTTPException`: 404 si el producto no está en la wishlist"
      operationId: remove_from_wishlist_api_v1_wishlist__product_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/wishlist/check/{product_id}:
    get:
      tags:
      - wishlist
      summary: Check Product In Wishlist
      description: "Verificar si un producto específico está en la lista de deseos\
        \ del usuario.\n\nArgs:\n    `product_id`: UUID del producto a verificar\n\
        \    `db`: Sesión de base de datos\n    `current_user`: Usuario autenticado\
        \ actual\n\nReturns:\n    `dict`: Indicador booleano de si el producto está\
        \ en la wishlist"
      operationId: check_product_in_wishlist_api_v1_wishlist_check__product_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
      description: Error de validación HTTP que contiene una lista de errores de validación
        detallados.
...
    ProductResponse-Output:
      properties:
        ean:
          type: string
          maxLength: 13
          minLength: 13
          title: Ean
        title:
          type: string
          title: Title
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        product_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Description
        category:
          type: string
          title: Category
        price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Price
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        discount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Discount
          default: 0
        rating:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Rating
          default: 0
        stock:
          type: integer
          minimum: 0
          title: Stock
        sponsored:
          type: boolean
          title: Sponsored
          default: false
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - ean
      - title
      - category
      - price
      - stock
      - id
      - created_at
      - updated_at
      title: ProductResponse
      description: Esquema de respuesta con la información completa de un producto,
        con campos numéricos serializados como strings para precisión decimal.
...
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
      - access_token
      - token_type
      title: Token
      description: Esquema de respuesta de autenticación con el token JWT de acceso
        y el tipo de token (bearer).
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        given_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Given Name
        paternal_surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Paternal Surname
        maternal_surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Maternal Surname
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - username
      - password
      title: UserCreate
      description: Esquema para crear un nuevo usuario en el sistema. Requiere email,
        nombre de usuario y contraseña. Los nombres son opcionales.
    UserLogin:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: UserLogin
      description: Esquema para iniciar sesión en el sistema. Requiere email y contraseña
        para autenticar al usuario.
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        given_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Given Name
        paternal_surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Paternal Surname
        maternal_surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Maternal Surname
        id:
          type: integer
          title: Id
        is_verified:
          type: boolean
          title: Is Verified
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - email
      - username
      - id
      - is_verified
      - created_at
      - updated_at
      title: UserResponse
      description: Esquema de respuesta con la información completa de un usuario,
        incluyendo ID, estado de verificación y fechas de creación/actualización.
    UserUpdate:
      properties:
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        given_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Given Name
        paternal_surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Paternal Surname
        maternal_surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Maternal Surname
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
      type: object
      title: UserUpdate
      description: Esquema para actualizar la información de un usuario. Todos los
        campos son opcionales, solo se actualizan los campos proporcionados.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
      description: Detalle de un error de validación individual que incluye la ubicación
        del error, el mensaje descriptivo y el tipo de error.
    WishlistCreate:
      properties:
        product_id:
          type: integer
          title: Product Id
      type: object
      required:
      - product_id
      title: WishlistCreate
      description: Esquema para agregar un producto a la lista de deseos del usuario.
        Requiere únicamente el ID del producto.
    WishlistList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/WishlistWithProduct-Output'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: WishlistList
      description: Respuesta con la lista completa de productos en la wishlist del
        usuario, incluyendo los detalles de cada producto y el total de items.
    WishlistResponse:
      properties:
        product_id:
          type: integer
          title: Product Id
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        added_at:
          type: string
          format: date-time
          title: Added At
      type: object
      required:
      - product_id
      - id
      - user_id
      - added_at
      title: WishlistResponse
      description: Esquema de respuesta con la información de un item en la wishlist,
        incluyendo el ID del producto, usuario y la fecha en que fue agregado.
...
    WishlistWithProduct-Output:
      properties:
        product_id:
          type: integer
          title: Product Id
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        added_at:
          type: string
          format: date-time
          title: Added At
        product:
          $ref: '#/components/schemas/ProductResponse-Output'
      type: object
      required:
      - product_id
      - id
      - user_id
      - added_at
      - product
      title: WishlistWithProduct
      description: Item de wishlist con la información completa del producto asociado,
        con campos numéricos serializados como strings.
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer